skills/hk-ipo-analyzer/scripts/requirements.txt
click>=8.1
rich>=13.0
requests>=2.31
beautifulsoup4>=4.12
pdfplumber>=0.10
matplotlib>=3.8
jinja2>=3.1
pyyaml>=6.0
